Set up requests page an commented css
There is now another html page in the forms app that is to view requests, i just set it up for basic navigation. The home page card now is what you have to click on to navigate but I cant figure out the text issues. Also working on making the images work I created a reference to the media directory and added the images there but it didnt work like it was suppsoed to so that is going to require some testing