Created a new Table and updated Old
Changed the names on the absence request table since I didnt realize the db automatically lowercased everything. Added a new table to manage the allowed days off for each day of the year. Currently it is only for 2024 so we will have to make it so it dynamically changes when the year changes or add more once it reaches a certain point of the year.