WIP: Fixed work order and working on pdfs
The issue with submitting work orders is now fixed and it works properly. Still working on jobs form, the navigation is not highlighting red, the create job posting is not formatted correctly. You can upload a pdf though and it displays the name on the job posting. Need to figure out how to create pdf